+++ server/user/src/main/java/com/yy/user/exception/AlreadyFollowedException.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.BAD_REQUEST)
 public class AlreadyFollowedException extends RuntimeException{
-    public AlreadyFollowedException(int userId, int followerUserId) {
+    public AlreadyFollowedException(long userId, long followerUserId) {
         super(String.format("%s is already followed by %s",userId,followerUserId));
     }
 }
